requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
django-widget-tweaks>=1.5
